src/weight-service/requirements.txt
paho-mqtt==1.6.1
kafka-python==2.0.2
requests==2.31.0